+++ rust/dbn/src/lib.rs
@@ -107,18 +107,3 @@ pub const UNDEF_STAT_QUANTITY: i32 = i32::MAX;
 pub const UNDEF_TIMESTAMP: u64 = u64::MAX;
 /// The length in bytes of the largest record type.
 pub const MAX_RECORD_LEN: usize = std::mem::size_of::<WithTsOut<v3::InstrumentDefMsg>>();
-
-/// Contains dataset code constants.
-#[deprecated(since = "0.25.0", note = "Use the `Dataset` enum instead")]
-pub mod datasets {
-    use crate::publishers::Dataset;
-
-    /// The dataset code for Databento Equities Basic.
-    pub const DBEQ_BASIC: &str = Dataset::DbeqBasic.as_str();
-    /// The dataset code for CME Globex MDP 3.0.
-    pub const GLBX_MDP3: &str = Dataset::GlbxMdp3.as_str();
-    /// The dataset code for OPRA PILLAR.
-    pub const OPRA_PILLAR: &str = Dataset::OpraPillar.as_str();
-    /// The dataset code for Nasdaq TotalView-ITCH.
-    pub const XNAS_ITCH: &str = Dataset::XnasItch.as_str();
-}